changed actors to floats

+++ src/Actor.cpp
@@ -1,6 +1,6 @@
 #include "Actor.h"
 
-Actor::Actor(int x, int y, int size) : m_playerShape(sf::Vector2f(size, size))
+Actor::Actor(float x, float y, float size) : m_playerShape(sf::Vector2f(size, size))
 {
 	mypos.m_x = x;
 	mypos.m_y = y;
@@ -14,7 +14,6 @@ sf::RectangleShape Actor::getShape() {
 	return m_playerShape;
 }
 
-
 void Actor::move(float x_dist, float y_dist) {
 	m_playerShape.move(x_dist, y_dist);
 }
@@ -25,7 +24,7 @@ void Actor::setTexture(std::string texturePath) {
 	m_playerShape.setTexture(&shapeTexture);
 }
 
-Player::Player(int x, int y, int size) : Actor(x, y, size)
+Player::Player(float x, float y, float size) : Actor(x, y, size)
 {
 
 }
